[Justice Counts] Add new_agency_this_week and new_state_this_week cols to CSG data pull (Recidiviz/recidiviz-data#24514)
## Description of the change

This PR adds 2 new columns to the CSG data pull:
1. `new_agency_this_week` - a boolean column that indicates if the
agency was created in the last 7 days
2. `new_state_this_week` - a boolean column that indicates if the agency
was created in the last 7 days AND it is associated with a
new`state_code`

How I Tested: I ran this locally on 10/10/23 and it produced the
following sheet. It correctly labeled "Cass County Justice Reinvestment
Advisory Council" as a new agency (created on 10/5/23). It correctly
labeled `new_state_this_week` as False for all agencies.

<img width="1840" alt="Screenshot 2023-10-10 at 5 02 19 PM"
src="https://github.com/Recidiviz/recidiviz-data/assets/82831800/7a9ca957-939c-4e26-80ef-75917545044c">

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@michelles-mbp.lan>
Co-authored-by: Michelle Orden <michelleorden@Michelles-MacBook-Pro.local>
GitOrigin-RevId: a41572da6d8495f246ce5bda745f3a297ce6496b