[US_TN] Update how annual reclass eligible folks are surfaced (Recidiviz/recidiviz-data#24603)
## Description of the change

In Recidiviz/recidiviz-data#23757 I updated the criteria query so that people are considered
eligible starting in the last week of the month prior to the month in
which they are eligible (i.e. if someone's reclass is due in November,
we want them to be surfaced starting in the last week of October).
However, because of `where` clause in
`us_tn_annual_reclassification_review_record` uses a `DATE_TRUNC` to
capture everyone eligible in the current month, it was also including
everyone eligible in the next month.

This PR adds an updated where clause which requires the current month to
match the month of the reclassification due date in order to surface
everyone eligible in the month, otherwise only surfacing if someone is
currently eligible.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b59d07ae18479f7a6020989f44a30c326aa17120