[US_MI] Update raw data config for RECIDIVIZ_REFERENCE_leadership_roster (Recidiviz/recidiviz-data#24680)
## Description of the change

Add a custom line terminator to the raw data config for
RECIDIVIZ_REFERENCE_leadership_roster so that the default dagger line
terminator won't need to be added to the manually-generated reference
table now or in the future.

Remove launch_env: is_local from
us_mi_RECIDIVIZ_REFERENCE_leadership_roster.yaml now that raw data can
be uploaded once this change is deployed.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#24660
Closes Recidiviz/recidiviz-data#24681

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: af8f6f7e3d4db8642705ee1649ba00f310cf315f