[Ingest] Temporarily remove the (very expensive) RunValidations step (Recidiviz/recidiviz-data#24734)
## Description of the change

The RunValidations step of ingest pipelines appears to contribute to
~90% of costs for some states (MO went from ~$100 to $12 without it, ME
went from ~$190 to $35 without it).

We think this is due to the amount of data that we are currently
shuffling around in this step to check invariants, which could be
optimized greatly.

For now, remove this step so we can run pipelines without the risk of
running a $200 pipeline by accident.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Prework for Recidiviz/recidiviz-data#24733 .

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 922dcdeefa5ef6a20fe86219755519c0bd1265e0