[US_TN] Update criteria to only consider Approved classifications (Recidiviz/recidiviz-data#24739)
## Description of the change

Based on feedback from some external validation, it appears that if a
classification is Denied (ClassificationDecision = 'D') then the annual
reclassification "clock" resets to use the previous available
classification. So I've updated the criteria query and completion event
query to only consider classifications where the decision is A
(approved).

The following query shows that this change increases # of people
eligible for an annual reclass by 16 and does not remove anyone as
eligible. This is expected since we're considering some previous
classifications "invalid" now so we'd expect more people to be
considered "due"

```
WITH main AS (
  select person_id
  from `recidiviz-123.dsharm20231016_main_task_eligibility_spans_us_tn.annual_reclassification_review_materialized`
  where '2023-10-01' between start_date and coalesce(end_date,'9999-01-01')
  and is_eligible
), sandbox_1 AS (
  select person_id
  from `recidiviz-123.dsharm20231016_2_task_eligibility_spans_us_tn.annual_reclassification_review_materialized`
  where '2023-10-01' between start_date and coalesce(end_date,'9999-01-01')
  and is_eligible

)
SELECT *
FROM main
FULL OUTER JOIN sandbox_1 USING(person_id)
WHERE main.person_id IS NULL AND sandbox_1.person_id IS NOT NULL
```

Two other notes:
- I filed Recidiviz/recidiviz-data#24737 as a TODO to bring ClassificationDecision into metadata
- I tried to rework the relevant views to bring
ClassificationDecisionDate from metadata but ran into entity deletion
issues (i.e. multiple CAF assessments for a single person on a single
day where 1 row has properly hydrated metadata and 1 row doesnt. When
ingest view results are rerun there is only 1 row). So I've updated the
TODO statement for that but not made any actual changes

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#24554

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4a34b8afe913c1d7b0e61ed3c7b5e2fc104f9426