[US_MI] Normalization revisions for MI Outliers (Recidiviz/recidiviz-data#24668)
## Description of the change

This PR:
- revises `infer_additional_periods` in incarceration periods
normalization to infer an additional incarceration period in the
following scenarios:
- If an incarceration period has admission reason `REVOCATION` or
`SANCTION_ADMISSION` and it wasn't immediately preceded by a supervision
period, infer an additional incarceration period between the most recent
supervision period and this incarceration period.
- If for an incarceration period the most recent supervision period
ended because of `REVOCATION` or `ADMITTED_TO_SUPERVISION` and this was
the first incarceration period after that supervision period, infer an
additional incarceration period between the most recent supervision
period and this incarceration period.
- revises `drop_bad_periods` to drop all `INVESTIGATION` periods since
we don't use investigation periods for anything and they are obscuring
the metrics.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Kris Gerig <kris@recidiviz.org>
GitOrigin-RevId: 047fa98cc3adfa1ceb8ab603a7d0640a5c40e0f7