[Justice Counts] Make sure prod deploy always allocates to 100% traffic (Recidiviz/recidiviz-data#24683)
## Description of the change

In our staging deploys, we always explicitly allocate a newly deployed
revision to 100% traffic. This is because, when we do a playtesting
deploy, we don't deploy to 100%. And for some reason, Cloud Run will
default to the amount of traffic from the last deployed revision. So if
the last deployed revision was a playtesting deploy to 0%, the next
staging deploy will also get deployed to 0%, unless we explicitly say
otherwise.

Since we don't do playtesting deploys to prod regularly, I didn't think
this was necessary to do in our prod script. However, I manually did a
playtesting deploy to prod the other day, and that meant that the next
prod deploy got 0%. So to be safe, I'm updating the prod deploy to do
the same thing as our staging deploy, and be explicit about making
traffic 100%!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ac987a4873dc61463e8413b2398e46dea34101ad