[Justice Counts] Script To Pull Agencies with Dashboards  (Recidiviz/recidiviz-data#24736)
## Description of the change

Script that will pull information about agency dashboards. It pulls the
number of capacity and cost metrics are reported, how many time periods
are reported, the agency dashboard link, and if the agency has a
dashboard enabled.
[This](https://docs.google.com/spreadsheets/d/1TcskbQat7a3OA7X-KRra74qBcXZMZVg5nOfdRFa9-64/edit#gid=1474106713)
is the output when run locally for production data!

I've confirmed manually that the information is correct for the agencies
that are in the spreadsheet.

## Related issues

Closes Recidiviz/recidiviz-data#23896

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cba63345564402183bde2c6dd1307bc8cb2a8b2f