[US_ME][JII] Update script to generate PDF packets directly (Recidiviz/recidiviz-data#24744)
## Description of the change

In https://github.com/Recidiviz/recidiviz-data/pull/24099, I created a
script that could be used to generate spreadsheets that could easily be
used to autofill Google Docs templates via an AppsScript that just does
some basic find/replace.
[Here](https://docs.google.com/spreadsheets/d/1iO5rHaVw6XQRq3uzWwjC8hP20TJZ0YUgrjoMG9bHRuA/edit#gid=428666279)
is an example of a spreadsheet it generated, and
[here](https://script.google.com/u/0/home/projects/1uPNgDFGyBfdJZq8F_5OOC_lxXUe8s-Ec0qLMKLa2z0ZLwXbkyvio9yY3/edit)
is the AppsScript that takes the spreadsheet and autofills the Google
Docs and generates the pdfs.

In this PR, I'm moving the logic from that AppsScript directly into this
script. So now, rather than having to generate the spreadsheet as an
intermediate step, we just directly generate the PDF packets.

There are a few steps now:
* Query BQ to get the raw data
* Format the data to extract exactly the fields needed to fill the
template
* Depending on the person's eligibility criteria, pick the right
template
* Make a copy of the template and put it in a folder
* Run find/replace
* Export the file as a pdf and put it in another folder
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: eb9903baca9a084b899a30925b56ce2049167231