Modifies task completion event source query to include all events (Recidiviz/recidiviz-data#24806)
## Description of the change

The TASK_COMPLETED event used to pull from a query looking at task
eligibility span ends -- this gives us an undercount of task
completions, since it requires that an eligibility span boundary exists
on the same date as a completion event. This change pulls directly from
`all_completion_events` since this is the superset of completion events,
including those that don't fall on a TES boundary. This required
modifying the `all_completion_events` output (which had no dependencies
prior to this PR) to ensure that the `task_type` field is a formatted
string name (it was previously outputting the string version of the enum
object, such that there was a `TaskCompletionEventType.` prefix on every
entry in this column).

Below is a query validating that aggregated metric outputs are now
aligned between the `task_completions_supervision_level_downgrades`
metric and the subset of `supervision_level_downgrades` that were not to
`limited`:
```
with sandbox as (
select
    state_code,
    start_date,
    task_completions_supervision_level_downgrade,
    supervision_level_downgrades - supervision_level_downgrades_to_limited as supervision_level_downgrades_no_limited,
from
    `recidiviz-staging.mayukas_24761_aggregated_metrics.supervision_state_aggregated_metrics_materialized`
where period = "YEAR"
and end_date = "2023-10-01"
)
,
staging as (
select
    state_code,
    start_date,
    task_completions_supervision_level_downgrade,
from
    `recidiviz-staging.aggregated_metrics.supervision_state_aggregated_metrics_materialized`
where period = "YEAR"
and end_date = "2023-10-01"
)

select
    state_code,
    staging.task_completions_supervision_level_downgrade as staging_task_completions,
    sandbox.task_completions_supervision_level_downgrade as sandbox_task_completions,
    sandbox.supervision_level_downgrades_no_limited
from staging join sandbox
using (state_code, start_date)
```

And below is a query checking that all supervision downgrade events
(excluding limited) derived from the `SUPERVISION_LEVEL_CHANGE` person
event are also present in the associated `TASK_COMPLETED` event:
```
with task_completions_staging as (
  select * from `recidiviz-staging.analyst_data.person_events_materialized`
  where event = "TASK_COMPLETED"
  and json_extract_scalar(event_attributes, "$.task_type") = "SUPERVISION_LEVEL_DOWNGRADE"
),
downgrades_staging as (
  select * from `recidiviz-staging.analyst_data.person_events_materialized`
  where event = "SUPERVISION_LEVEL_CHANGE"
  and json_extract_scalar(event_attributes, "$.change_type") = "DOWNGRADE"
  and json_extract_scalar(event_attributes, "$.new_supervision_level") != "LIMITED"
)
,
task_completions_sandbox as (
  select * from `recidiviz-staging.mayukas_24761_analyst_data.person_events_materialized`
  where event = "TASK_COMPLETED"
  and json_extract_scalar(event_attributes, "$.task_type") = "SUPERVISION_LEVEL_DOWNGRADE"

),
downgrades_sandbox as (
  select * from `recidiviz-staging.mayukas_24761_analyst_data.person_events_materialized`
  where event = "SUPERVISION_LEVEL_CHANGE"
  and json_extract_scalar(event_attributes, "$.change_type") = "DOWNGRADE"
  and json_extract_scalar(event_attributes, "$.new_supervision_level") != "LIMITED"
)
select * from (
  select downgrades_staging.state_code, countif(task_completions_staging.person_id is null) as staging_error from downgrades_staging left join task_completions_staging using (person_id, event_date) group by 1
)
join (
  select downgrades_sandbox.state_code, countif(task_completions_sandbox.person_id is null) as sandbox_error from downgrades_sandbox left join task_completions_sandbox using (person_id, event_date) group by 1
)
using (state_code)
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#24761

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 71916daa5e3014eeaafe5b9b120cb6ec9241e610